app/requirements.txt
fastapi
uvicorn[standard]

python-dotenv
pydantic

python-jose[cryptography]
passlib[bcrypt]

python-multipart

pytesseract
Pillow
easyocr

openai-whisper

sentence-transformers
numpy
scikit-learn

faiss-cpu
chromadb

openai
tiktoken

sqlalchemy
psycopg2-binary

requests
loguru

pytest
httpx

langchain
langchain-community
langchain-openai
langchain_huggingface
chromadb
PyPDF2
langchain_chroma
PyYAML==6.0.2